Move decorator location.
